refactor: 직업 공고 목록과 직업 공고 상세 조회 기능 분리
클래스를 별도로 생성하여 직업 공고 리스트를 불러오는 기능과 직업 공고 상제 조회 기능을 분리
